Use /sys/block/*/stat instead of /proc/diskstats
/proc/diskstats includes both disks and partitions, relaying on
collector.diskstats.ignored-devices to filter them out. However, the
device name might be deceiving like xvda1 (xen) that can both represent
a partition in xvda or a standalone disk.

A new config collector.diskstats.include-partitions was added to include
or not partitions. It was kept on by default to maintain the previous
behavior.

The diskstats_linux_test.go was not implemented for /sys/block.

Signed-off-by: Luiz Angelo Daros de Luca <luizluca@gmail.com>